Took inner classes out of TreeviewController
Makes TreeItem and Cell classes easier to reuse. Also reduces clutter,
in preparation of TVC filter fixes
